+++ plugins/animation_utils/src/codec.ts
@@ -52,6 +52,28 @@ function onProjectParse(e: any) {
 }
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 function onBedrockCompile(e: any) {
+    if (Format.id !== "animated_entity_model") return;
+
+    // Remove display transforms from non-bedrock geometry
+    const geometry = e.model["minecraft:geometry"];
+
+    if (geometry) {
+        geometry.forEach((geo: Map<string, any>) => {
+            delete geo["item_display_transforms"];
+        })
+    }
+
+    // Force-suppress specific version advancement for non-bedrock models to prevent legacy version crashes until a better system is established
+    switch (e.model.format_version) {
+        case "1.14.0":
+        case "1.21.0":
+        case "1.21.20":
+            e.model.format_version = "1.12.0"
+            break;
+        default:
+            break;
+    }
+
     // console.log('onBedrockCompile e:', e);
     // maybeExportItemJson(e.options);
 }
@@ -65,6 +87,53 @@ function animatorBuildFile() {
                 'geckolib_format_version': geckoSettings.formatVersion,
             }
         );
+
+        // Convert exported bedrock animations to non-bedrock
+        // This should be a menu item but that can be a future thing
+        if (res.animations) {
+            for (const animation in res.animations) {
+                const bones = res.animations[animation].bones;
+
+                if (bones) {
+                    for (const boneName in bones) {
+                        const bone = bones[boneName];
+
+                        for (const animationGroupType in bone) {
+                            const animationGroup = bone[animationGroupType];
+
+                            for (const timestamp in animationGroup) {
+                                const keyframe = animationGroup[timestamp];
+
+                                if (!keyframe)
+                                    continue
+
+                                let bedrockKeyframe : Map<any, any> = keyframe.pre;
+                                let bedrockKeyframeData : Map<any, any> = undefined;
+
+                                if (bedrockKeyframe !== undefined) {
+                                    bedrockKeyframeData = bedrockKeyframe;
+                                    delete keyframe.pre
+                                }
+
+                                bedrockKeyframe = keyframe.post;
+
+                                if (bedrockKeyframe !== undefined) {
+                                    bedrockKeyframeData = bedrockKeyframe;
+                                    delete keyframe.post
+                                }
+
+                                if (bedrockKeyframeData !== undefined) {
+                                    Object.assign(keyframe, bedrockKeyframeData)
+
+                                    if (keyframe.lerp_mode)
+                                        delete keyframe.lerp_mode;
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
     }
     // console.log('animatorBuildFile res:', res);
     return res;
@@ -97,6 +166,24 @@ function getKeyframeDataPoints(source: any) {
     }
 }
 
+function geoLoopToBbLoop(jsonLoop) {
+    if (jsonLoop) {
+        if (typeof jsonLoop === 'boolean') {
+            return jsonLoop ? 'loop' : 'once'
+        }
+
+        if (typeof jsonLoop === 'string') {
+            if (jsonLoop === "hold_on_last_frame")
+                return 'hold'
+
+            if (jsonLoop === "loop" || jsonLoop === "true")
+                return 'loop'
+        }
+    }
+
+    return 'once'
+}
+
 function animatorLoadFile(file, animation_filter) {
     // Currently no modifications are needed
     // eslint-disable-next-line no-undef
@@ -111,9 +198,7 @@ function animatorLoadFile(file, animation_filter) {
             const animation = new Blockbench.Animation({
                 name: ani_name,
                 path,
-                // TODO: Make sure it's OK to disable this line
-                // loop: a.loop && (a.loop == 'hold_on_last_frame' ? 'hold' : 'loop'),
-                loop: a.loop,
+                loop: geoLoopToBbLoop(a.loop),
                 override: a.override_previous_animation,
                 anim_time_update: (typeof a.anim_time_update == 'string'
                     ? a.anim_time_update.replace(/;(?!$)/, ';\n')
@@ -274,6 +359,22 @@ export function maybeExportItemJson(options = {}) {
             blockmodel.display = new_display
         }
     }
+    if (checkExport('textures', Object.keys(Project.textures).length >= 1)) {
+        for (const tex of Project.textures) {
+            if (tex.particle || Object.keys(Project.textures).length === 1) {
+                let name = tex.name;
+
+                if (name.indexOf(".png") > 0)
+                    name = name.substring(0, name.indexOf(".png"))
+
+                const texturesMap = {}
+                texturesMap['particle'] = name
+                blockmodel.textures = texturesMap;
+
+                break
+            }
+        }
+    }
 
     const blockmodelString = JSON.stringify(blockmodel, null, 2);
 
@@ -310,6 +411,7 @@ const format = new ModelFormat({
     bone_rig: true,
     centered_grid: true,
     animated_textures: true,
+    select_texture_for_particles: true,
     animation_mode: true,
     animation_files: true,
     locators: true,